move down the partner list state
